requirements_full.txt
numpy==1.24.1
opencv_python==4.7.0.72
Pillow==9.4.0
plotly==5.13.1
scipy==1.9.1
streamlit==1.18.1
